decouple the exception mediation
now you can inject another mediator, different from the default one
