+++ src/main/java/com/github/novotnyr/springframework/web/jsonschema/JsonSchemaResolver.java
@@ -10,7 +10,8 @@ import org.springframework.web.context.request.NativeWebRequest;
 public interface JsonSchemaResolver {
 
     /**
-     * Resolves json schemas according to method parameter and webrequest.
+     * Resolves json schemas according to method parameter and webrequest. Must always return a resource, if
+     * it is unavailable you can send a not found Resource
      */
     Resource resolveJsonSchemaResource(MethodParameter methodParameter, NativeWebRequest webRequest);
 }
